Remove invalid export from route
