+++ app/api/auth/[...nextauth]/route.ts
@@ -1,15 +1,5 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID as string,
-      clientSecret: process.env.GOOGLE_SECRET_ID as string,
-    }),
-  ],
-  secret: process.env.NEXTAUTH_SECRET as string,
-};
+import { authOptions } from '@/utils/auth';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
